requirements.txt
bme680==2.0.0
bmm150==0.2.2
enum-tools==0.9.0.post1
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
Pygments==2.18.0
pyserial==3.5
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
Seeed-grove.py==0.8
seeed-python-ircamera==1.9.3
seeed-python-mlx9064x==1.1.4
sgp30==0.1.6
smbus2==0.4.3
typing_extensions==4.12.2
pigpio==1.78